fix: app stopwatch
Update resetBoard() method in app.js script to use the
`destroy` method instead of the reset method. The `reset`
public method of the stopwatch was not correctly resetting the
app stop watch when the game reset. This update simply destroys it
so the stopwatch can simply be re-initiated inside the `newGame()`
method.