fix: jittery stopwatch timer in app's score-panel
## Update app.css styles

Add new `.timer` class to make the font of the stopwatch timer a
monospace font family to prevent shaking seen in sans-serif font.

Adjust font-size to work with new font family specified.

## Update stopwatch.js

The `createTimerEl` method adds the value `timer` so that the
`timer` css class can be applied to fix shaky stopwatch timer.